+++ src/main/java/github/anandb/netbeans/ui/IconPreviewManager.java
@@ -3,7 +3,6 @@ package github.anandb.netbeans.ui;
 import java.awt.Dimension;
 import java.awt.Image;
 import java.io.File;
-import javax.swing.BorderFactory;
 import javax.swing.ImageIcon;
 import javax.swing.JLabel;
 import javax.swing.JMenuItem;
@@ -38,7 +37,6 @@ final class IconPreviewManager {
 
     private void initPreviewLabel() {
         previewLabel.setPreferredSize(new Dimension(80, 80));
-        previewLabel.setBorder(BorderFactory.createLineBorder(ThemeManager.getCurrentTheme().mutedForeground()));
         previewLabel.setHorizontalAlignment(SwingConstants.CENTER);
         previewLabel.setToolTipText("Right-click to clear icon");
         previewLabel.addMouseListener(new MouseAdapter() {